+++ test/findBin.spec.js
@@ -9,15 +9,17 @@ const packageJSON = {
     },
     'lint-staged': {}
 }
-const npmWichMockGood = (path, cb) => {
-    cb(null, path)
+const npmWichMockGood = {
+    sync: path => path
 }
-const npmWichMockBad = (path, cb) => {
-    cb(true, null)
+const npmWichMockBad = {
+    sync: path => {
+        throw new Error('not found: ' + path)
+    }
 }
 
 describe('findBin', () => {
-    it('should return npm run command if it exist in both package.json and .bin/', done => {
+    it('should return npm run command if it exist in both package.json and .bin/', () => {
         const packageJSONMock = {
             scripts: {
                 eslint: 'eslint'
@@ -25,40 +27,38 @@ describe('findBin', () => {
         }
 
         findBin.__set__('npmWhich', npmWichMockGood)
-        findBin('eslint', 'test.js', packageJSONMock, (err, bin, args) => {
-            expect(err).toBe(null)
-            expect(bin).toEqual('npm')
-            expect(args).toEqual(['run', '-s', 'eslint', '--', 'test.js'])
-            done()
-        })
+        const { bin, args } = findBin('eslint', 'test.js', packageJSONMock)
+        expect(bin).toEqual('npm')
+        expect(args).toEqual(['run', '--silent', 'eslint', '--', 'test.js'])
     })
 
-    it('should return bin from node_modules/.bin if there is no command in package.json', done => {
+    it('should return bin from node_modules/.bin if there is no command in package.json', () => {
         findBin.__set__('npmWhich', npmWichMockGood)
-        findBin('eslint', 'test.js test2.js', packageJSON, (err, bin, args) => {
-            expect(err).toBe(null)
-            expect(bin).toEqual('eslint')
-            expect(args).toEqual(['--', 'test.js test2.js'])
-            done()
-        })
+        const { bin, args } = findBin('eslint', 'test.js test2.js', packageJSON)
+        expect(bin).toEqual('eslint')
+        expect(args).toEqual(['--', 'test.js test2.js'])
     })
 
-    it('should return bin from node_modules/.bin on missed scripts in package.json', done => {
+    it('should parse cmd and add arguments to args', () => {
         findBin.__set__('npmWhich', npmWichMockGood)
-        findBin('eslint', 'test.js test2.js', { }, (err, bin, args) => {
-            expect(err).toBe(null)
-            expect(bin).toEqual('eslint')
-            expect(args).toEqual(['--', 'test.js test2.js'])
-            done()
-        })
+        const { bin, args } = findBin('eslint --fix', 'test.js test2.js', packageJSON)
+        expect(bin).toEqual('eslint')
+        expect(args).toEqual(['--fix', '--', 'test.js test2.js'])
+    })
+
+    it('should return bin from $PATH if there is no command in package.json and no bin in node_modules', () => {
+        findBin.__set__('npmWhich', npmWichMockBad)
+        findBin.__set__('which', npmWichMockGood)
+        const { bin, args } = findBin('git add', 'test.js test2.js', packageJSON)
+        expect(bin).toEqual('git')
+        expect(args).toEqual(['add', '--', 'test.js test2.js'])
     })
 
     it('should return error if bin not found and there is no entry in scripts section', () => {
         findBin.__set__('npmWhich', npmWichMockBad)
+        findBin.__set__('which', npmWichMockBad)
         expect(() => {
-            findBin('eslint', 'test.js', packageJSON, (err) => {
-                throw new Error(err)
-            })
-        }).toThrow()
+            findBin('eslint', 'test.js', packageJSON)
+        }).toThrow('eslint could not be found. Try `npm install eslint`.')
     })
 })